requirements.txt
beautifulsoup4==4.10.0
Pillow==9.0.1
python_telegram_bot==13.11
requests==2.25.1
telegram==0.0.1
lxml==4.7.1
